#2144 roadway_objects: file renaming
